requirements.txt
astroid==2.15.8
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
colorama==0.4.6
dill==0.3.7
gensim==4.3.2
idna==3.4
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
numpy==1.26.0
platformdirs==3.10.0
pycparser==2.21
pydub==0.25.1
pylint==2.17.7
python-rake==1.5.0
requests==2.31.0
scipy==1.11.3
six==1.16.0
smart-open==6.4.0
srt==3.5.3
tomlkit==0.12.1
tqdm==4.66.1
urllib3==2.0.5
vosk==0.3.45
websockets==11.0.3
wrapt==1.15.0
